Rename and move MPA RTC minplus to match current DNC implementation naming and location
